fix menu

+++ database/migrations/2025_10_30_061656_create_menus_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->string('name');
             $table->foreignId('category_id')->constrained()->cascadeOnDelete();
             // $table->enum('category', ['coffee', 'non-coffee'])->nullable();
-            $table->decimal('base_price', 10, 2); // harga dasar (biasanya hot)
+            $table->bigInteger('base_price');
             $table->text('description')->nullable();
             $table->string('photo_path')->nullable();
             $table->enum('status', ['available', 'unavailable'])->default('available');
